Adding the Assurance Banner
- Added assurance banner across entire app
- Some cleanup and refactoring of file naming scheme
- Added TOD Calculator placeholder
