+++ Sample/Kitchen Sink/Demos/CrossWindViewModel.swift
