+++ Sample/Kitchen Sink/Demos/TODCalculatorViewModel.swift
@@ -1,0 +1,12 @@
+//
+//  TODCalculatorViewModel.swift
+//  flight-ui-ios
+//
+//  Created by Appivate 2024
+//
+
+import Foundation
+
+class TODCalculatorViewModel: ObservableObject {
+    
+}
